fix: move acp permission helper out of component file
