+++ src/web-ui/src/flow_chat/tool-cards/AcpPermissionActions.utils.ts
@@ -1,0 +1,5 @@
+import type { FlowToolItem } from '../types/flow-chat';
+
+export function hasAcpPermissionOptions(toolItem: FlowToolItem): boolean {
+  return Boolean(toolItem.acpPermission?.options?.length);
+}
